style: moved query string util to static class
